+++ CRUDRecipeEF.PL/appsettings.json
@@ -1,0 +1,5 @@
+{
+  "ConnectionStrings": {
+    "Default": "Data Source = Recipe.db"
+  }
+}
